Dvx.Interpreter

Description

Documentation

execute

Arguments

∷ [Dvx.Parser.Context]
Current context chain (initial context when first called)

→ [Dvx.Parser.DvxExpr]
List of expressions to evaluate

→ GHC.Types.IO ()

Executes a list of expressions

executeExpr

Arguments

∷ [Dvx.Parser.Context]
Current context chain

→ Dvx.Parser.DvxExpr
Expression to evaluate

→ GHC.Types.IO (Dvx.Parser.DvxValue, [Dvx.Parser.Context])

Executes a single expression, returning a value and updated context |Throws an error if the expression can't be evaluated

appendContext

Arguments

∷ [Dvx.Parser.Context]
Context to append value to

→ GHC.Base.String
Name of the value to append

→ Dvx.Parser.DvxValue
Value to append

→ [Dvx.Parser.Context]
Resulting context

Adds a value (variable, function def) to the head of a context chain

getVar

Arguments

∷ [Dvx.Parser.Context]
Context chain to search variable from

→ GHC.Base.String
Name to search for

→ Dvx.Parser.DvxValue
Resulting value

Gets a variable from the context chain |Throws an error if it can't find the variable

setVar

Arguments

∷ [Dvx.Parser.Context]
Context chain to add variable to

→ GHC.Base.String
Name of the variable to add

→ Dvx.Parser.DvxExpr
Expresion of the variable to add (resolved in place)

→ GHC.Types.IO (Dvx.Parser.DvxValue, [Dvx.Parser.Context])

Sets a variable in the context chain

makeFunction

Arguments

∷ [GHC.Base.String]
Argument list

→ Dvx.Parser.DvxExpr
Expression body

→ Dvx.Parser.Function
Resulting function

Creates a function from a list of arguments and a return expression

bindArgs

Arguments

∷ [GHC.Base.String]
Argument list

→ [Dvx.Parser.DvxValue]
Value list

→ Dvx.Parser.Context
Resulting context

Creates a function context based from a list of arguments and their values

apply

Arguments

∷ [Dvx.Parser.Context]
Context chain

→ Dvx.Parser.DvxValue
Function resolved from the expression

→ [Dvx.Parser.DvxValue]
List of arguments applied to the function

→ GHC.Types.IO Dvx.Parser.DvxValue
Function's return value

Executes an expression where possible, otherwise returns nil

resolve

Arguments

∷ [Dvx.Parser.Context]
Context chain to resolve from

→ [Dvx.Parser.DvxExpr]
Expressions to resolve

→ GHC.Types.IO [Dvx.Parser.DvxValue]
Resolved values

Retrieves a list of values from a list of expressions |This is just a wrapper over resolveValue

resolveValue

Arguments

∷ [Dvx.Parser.Context]
Context chain to resolve from

→ Dvx.Parser.DvxExpr
Expression to resolve

→ GHC.Types.IO Dvx.Parser.DvxValue
Resolved value

Retrieves a value from an expression |Works by executing the expression where possible or just looking around in |the context chain.