Dvx.Parser

Description

Documentation

type Context = [(GHC.Base.String, Dvx.Parser.DvxValue)]

type Function = [Dvx.Parser.Context] → [Dvx.Parser.DvxValue] → GHC.Types.IO Dvx.Parser.DvxValue

data DvxValue

Constructors

TypeNil

TypeBool GHC.Types.Bool

TypeInt GHC.Types.Int

TypeStr GHC.Base.String

TypeFun Dvx.Parser.Function

TypeLst [Dvx.Parser.DvxValue]

data DvxExpr

Constructors

DvxTok Dvx.Tokens.DvxToken

DvxStart

DvxConst Dvx.Parser.DvxValue

DvxVar GHC.Base.String

DvxCall GHC.Base.String [Dvx.Parser.DvxExpr]

DvxFunc GHC.Base.String [GHC.Base.String] Dvx.Parser.DvxExpr

DvxDecl GHC.Base.String Dvx.Parser.DvxExpr

DvxList [Dvx.Parser.DvxExpr]

separators ∷ GHC.Base.String

parse ∷ [Dvx.Tokens.DvxToken] → [Dvx.Parser.DvxExpr]

makeast ∷ [Dvx.Parser.DvxExpr] → [Dvx.Parser.DvxExpr]

Makes a fully parsed AST off a raw tree

discover ∷ Dvx.Tokens.DvxToken → Dvx.Parser.DvxExpr

Parses plain tokens (DvxTok) into something meaningful

getArgs ∷ [Dvx.Parser.DvxExpr] → [GHC.Base.String]

Gets a list of strings from a function declaration argument DvxList

parseTokens ∷ Dvx.Tokens.DvxToken → [Dvx.Parser.DvxExpr] → [Dvx.Parser.DvxExpr]

Parses a list plain tokens (DvxTok) into a tree

prependList ∷ Dvx.Parser.DvxExpr → Dvx.Tokens.DvxToken → Dvx.Parser.DvxExpr

Adds token to the beginning of a list, if both arguments are token, create a list of both

parseValue ∷ GHC.Base.String → Dvx.Tokens.DvxToken

Given a String, returns the corresponding token with its semantic value, if any.

tokenize ∷ [GHC.Base.String] → [Dvx.Tokens.DvxToken]

Creates a list of tokens off lines of code