Dvx.Utils

Description

Documentation

trim ∷ GHC.Base.String → GHC.Base.String

Trim whitespace from a string

splitOn ∷ GHC.Classes.Eq a ⇒ a → [a] → [[a]]

Split a list of lists based on a single item separator

splitAndKeep ∷ GHC.Base.String → GHC.Base.String → [GHC.Base.String]

Split a string based on multiple separators while keeping them as items

isNumeric ∷ GHC.Base.String → GHC.Types.Bool

Check if a string is an integer

middle ∷ GHC.Base.String → GHC.Base.String

Take out the first and last character of a string

joinsub ∷ [[a]] → [a]

Merge a list of lists into a single list

joinstr ∷ [GHC.Base.String] → GHC.Types.Int → [GHC.Base.String] → [GHC.Base.String]

Create string tokens from a list of strings This is a workaround to a design fault in our tokenizer

boolToInt ∷ GHC.Types.Bool → GHC.Types.Int

1 if given True, 0 if given False